refactor(core): move chain API types into core and introduce TipValidationApi
Extract shared chain API/domain types from `chain` into `core` so they can be
consumed without depending on chain implementation packages.

- move `BaseBlockChain`, `Blockchain`, `BlockchainUpdate`, and `ConnectTipResult`
  to `core/src/main/scala/org/bitcoins/core/api/chain`
- move `TipUpdateResult` to `core/api/chain`
- add `TipValidationApi` in `core/api/chain`
- remove obsolete `CheckHeaderResult` from `chain` package
