--- chain/src/main/scala/org/bitcoins/chain/blockchain/CheckHeaderResult.scala
@@ -1,5 +1,0 @@
-package org.bitcoins.chain.blockchain
-
-import org.bitcoins.chain.validation.TipUpdateResult
-
-case class CheckHeaderResult(result: TipUpdateResult, chain: Blockchain)
