+++ chain/src/main/scala/org/bitcoins/chain/validation/TipValidation.scala
@@ -1,8 +1,13 @@
 package org.bitcoins.chain.validation
 
 import org.bitcoins.chain.ChainVerificationLogger
-import org.bitcoins.chain.blockchain.Blockchain
 import org.bitcoins.chain.pow.Pow
+import org.bitcoins.core.api.chain.{
+  Blockchain,
+  ConnectTipResult,
+  TipUpdateResult,
+  TipValidationApi
+}
 import org.bitcoins.core.api.chain.db.{BlockHeaderDb, BlockHeaderDbHelper}
 import org.bitcoins.core.number.UInt32
 import org.bitcoins.core.protocol.blockchain.{BitcoinChainParams, BlockHeader}
@@ -12,7 +17,9 @@ import org.bitcoins.core.util.NumberUtil
   * blockchain. The checks things like proof of work difficulty, if it
   * references the previous block header correctly etc.
   */
-sealed abstract class TipValidation extends ChainVerificationLogger {
+sealed abstract class TipValidation
+    extends TipValidationApi
+    with ChainVerificationLogger {
 
   /** Checks if the given header can be connected to the current tip This is the
     * method where a
@@ -65,6 +72,89 @@ sealed abstract class TipValidation extends ChainVerificationLogger {
 
     logTipResult(connectTipResult, currentTip)
     connectTipResult
+  }
+
+  override def connectTip(
+      header: BlockHeader,
+      blockchain: Blockchain,
+      chainParams: BitcoinChainParams): ConnectTipResult = {
+    logger.debug(
+      s"Attempting to add new tip=${header.hashBE.hex} with prevhash=${header.previousBlockHashBE.hex} to chain"
+    )
+
+    val tipResult: ConnectTipResult = {
+      findPrevBlockHeaderIdx(header, blockchain) match {
+        case None =>
+          logger.debug(
+            s"No common ancestor found in the chain with tip=${blockchain.tip.hashBE.hex} to connect to hash=${header.hashBE.hex} prevHash=${header.previousBlockHashBE.hex}. This may be because we have a competing reorg!"
+          )
+          val err = TipUpdateResult.BadPreviousBlockHash(header)
+          val failed = ConnectTipResult.BadTip(err)
+          failed
+
+        case Some(prevHeaderIdx) =>
+          // found a header to connect to!
+          val prevBlockHeader = blockchain.headers(prevHeaderIdx)
+          logger.debug(
+            s"Attempting to add new tip=${header.hashBE.hex} with prevhash=${header.previousBlockHashBE.hex} to chain of ${blockchain.length} headers with tip ${blockchain.tip.hashBE.hex}"
+          )
+          val chain = blockchain.fromValidHeader(prevBlockHeader)
+          val tipResult =
+            TipValidation.checkNewTip(newPotentialTip = header,
+                                      chain,
+                                      chainParams)
+
+          tipResult match {
+            case success: TipUpdateResult.Success =>
+              logger.debug(
+                s"Successfully verified=${success.header.hashBE.hex}, connecting to chain"
+              )
+              val connectionIdx = blockchain.length - prevHeaderIdx
+
+              // we construct a new blockchain by prepending the headers vector from the old one with the new tip
+              // in order to avoid creating unnecessary hidden copies of the blockchain here
+              if (connectionIdx != blockchain.length) {
+                val newChain = Blockchain(
+                  success.headerDb +: blockchain.headers.takeRight(
+                    connectionIdx
+                  )
+                )
+                // means we have a reorg, since we aren't connecting to latest tip
+                ConnectTipResult.Reorg(success, newChain)
+              } else {
+                val olderChain = if (blockchain.size < 2016) {
+                  blockchain.headers
+                } else blockchain.headers.take(2015)
+                val newChain = Blockchain(success.headerDb +: olderChain)
+                // we just extended the latest tip
+                ConnectTipResult.ExtendChain(success, newChain)
+              }
+            case fail: TipUpdateResult.Failure =>
+              logger.warn(
+                s"Could not verify header=${header.hashBE.hex}, reason=$fail"
+              )
+              ConnectTipResult.BadTip(fail)
+          }
+      }
+    }
+    tipResult
+  }
+
+  /** Finds the parent's index of the given header
+    */
+  private def findPrevBlockHeaderIdx(
+      header: BlockHeader,
+      blockchain: Blockchain
+  ): Option[Int] = {
+    // Let's see if we are lucky and the latest tip is the parent.
+    val latestTip = blockchain.tip
+    if (latestTip.hashBE == header.previousBlockHashBE) {
+      // Yes we are.
+      Some(0)
+    } else {
+      // No. Scanning the blockchain to find the parent.
+      blockchain.findHeaderIdx(header.previousBlockHashBE)
+    }
   }
 
   /** Logs the result of
